Ultimos cambios
Subida cambios

+++ VS_YEngine/Vector.cpp
@@ -5,8 +5,8 @@ vec3_t Vector::vec3_rotate_x(vec3_t v, float angle)
 {
 	vec3_t _v{
 		.x = v.x,
-		.y = v.y * cos(angle) - v.z * sin(angle),
-		.z = v.y * sin(angle) + v.z * cos(angle)
+		.y = v.y * static_cast<float>(cos(angle)) - v.z * static_cast<float>(sin(angle)),
+		.z = v.y * static_cast<float>(sin(angle)) + v.z * static_cast<float>(cos(angle))
 	};
 
 	return _v;
@@ -15,9 +15,9 @@ vec3_t Vector::vec3_rotate_x(vec3_t v, float angle)
 vec3_t Vector::vec3_rotate_y(vec3_t v, float angle)
 {
 	vec3_t _v{
-		.x = v.x * cos(angle) - v.z * sin(angle),
+		.x = v.x * static_cast<float>(cos(angle)) - v.z * static_cast<float>(sin(angle)),
 		.y = v.y,
-		.z = v.x * cos(angle) + v.z * cos(angle)
+		.z = v.x * static_cast<float>(cos(angle)) + v.z * static_cast<float>(cos(angle))
 	};
 
 	return _v;
@@ -26,8 +26,8 @@ vec3_t Vector::vec3_rotate_y(vec3_t v, float angle)
 vec3_t Vector::vec3_rotate_z(vec3_t v, float angle)
 {
 	vec3_t _v{
-		.x = v.x * cos(angle) - v.y * sin(angle),
-		.y = v.x * sin(angle) + v.y * cos(angle),
+		.x = v.x * static_cast<float>(cos(angle)) - v.y * static_cast<float>(sin(angle)),
+		.y = v.x * static_cast<float>(sin(angle)) + v.y * static_cast<float>(cos(angle)),
 		.z = v.z
 	};
